Generate /tofu/series.json metadata
This new metadata file summarizes our minor release series rather than
describing all of the individual release versions separately as in
api.json.

This is mainly for our own use to drive other automated processes that need
to work on a per-series basis, like our automated scanning for upstream
security vulnerabilities in the main opentofu repository, but it might
also be useful for others who want to easily track the latest release in
a particular series, or know when the series they are using has become
end-of-life, or similar.

Signed-off-by: Martin Atkins <mart@degeneration.co.uk>